Update GTN organization Bioschemas markup
Work from ELIXIR BioHackathon 2025, project 18 "Mining the potential of knowledge graphs". We are working on an MCP server which will allow people to query multiple training registries such as GTN using LLMs. The knowledge graph we build will be much richer if we can include identifiers for people and organisations.

We noticed that the Bioschemas Organization profile 0.2 you use in GTN training materials to describe the GTN itself has problems. It uses the incorrect `conformsTo` type. It is missing the required property `description`. It uses the outdated version 0.2. All three of these are correct in the other places you use the Organization Bioschemas profile (when you describe the orgs that trainers are members of).  

I have proposed corrections/updates for the three matters. (There are further required properties missing but I am going to the Bioschemas community to propose a version 0.4 draft profile which makes these properties recommended instead of required. I am also looking to use ROR as the recommended identifier.)